Updated interface
The VehicleBanChecker interface used to require the Vehicle. This has had unintended consequences when integreting with the freight contrib where the Vehicles are not in the scenario, but the CarrierCapabilities. Using the Id<Vehicle> circumvents the issue as the user need to decide, when implementing the VehicleBanChecker, WHERE to look forn the vehicle.